refactor: useShare 훅 분리
